feat: implement task details & refactor ids to long
- Refactor `Task` and `SubTask` identifiers from `Int` to `Long` across Domain, Data, and UI layers.
- Implement `TaskDetailScreen` and `TaskDetailViewModel` to display task info and manage subtasks.
- Add functionality to toggle subtask completion status within the detail view.
- Enable navigation from `TaskListScreen` to `TaskDetailScreen` using `TaskDetailRoute`.
- Implement delete task functionality in `TaskListViewModel` and expose it to the UI.
- Update `TaskMapper` and `SubTaskMapper` to explicitly map IDs when converting models to entities.
- Configure `AppNavHost` to handle `TaskDetailRoute` and pass navigation callbacks.

+++ app/src/main/java/com/ab/todo/navigation/Routes.kt
@@ -9,4 +9,4 @@ data object TaskListRoute
 data object AddTaskRoute
 
 @Serializable
-data class TaskDetailRoute(val id: Int)
+data class TaskDetailRoute(val id: Long)
