+++ data/src/main/java/com/ab/data/datasource/subtask/SubTaskLocalDataSource.kt
@@ -4,8 +4,8 @@ import com.ab.data.models.SubTaskEntity
 import kotlinx.coroutines.flow.Flow
 
 interface SubTaskLocalDataSource {
-    suspend fun listAllByTaskId(taskId: Int): List<SubTaskEntity>
-    fun listAllByTaskIdAsFlow(taskId: Int): Flow<List<SubTaskEntity>>
+    suspend fun listAllByTaskId(taskId: Long): List<SubTaskEntity>
+    fun listAllByTaskIdAsFlow(taskId: Long): Flow<List<SubTaskEntity>>
     suspend fun insert(subTask: SubTaskEntity)
     suspend fun insert(subTasks: List<SubTaskEntity>)
     suspend fun update(subTask: SubTaskEntity)
