+++ domain/src/main/java/com/ab/domain/respository/SubTaskRepository.kt
@@ -4,8 +4,8 @@ import com.ab.domain.model.SubTaskModel
 import kotlinx.coroutines.flow.Flow
 
 interface SubTaskRepository {
-    suspend fun listByTaskId(id: Int): List<SubTaskModel>
-    fun listByTaskIdAsFlow(id: Int): Flow<List<SubTaskModel>>
+    suspend fun listByTaskId(id: Long): List<SubTaskModel>
+    fun listByTaskIdAsFlow(id: Long): Flow<List<SubTaskModel>>
     suspend fun add(task: SubTaskModel)
     suspend fun addList(tasks: List<SubTaskModel>)
     suspend fun edit(task: SubTaskModel)
