+++ domain/src/main/java/com/ab/domain/respository/TaskRepository.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 interface TaskRepository {
     suspend fun listAll(): List<TaskModel>
     fun listAllAsFlow(): Flow<List<TaskModel>>
-    suspend fun getById(id: Int): TaskModel?
+    suspend fun getById(id: Long): TaskModel?
     suspend fun add(task: TaskModel): Long
     suspend fun edit(task: TaskModel)
     suspend fun delete(task: TaskModel)
